PDFBOX-4883: replace doubleFloat() with floatValue()
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1879274 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/pdfbox/preflight/content/StubOperator.java
@@ -255,7 +255,8 @@ public class StubOperator extends OperatorProcessor
             }
 
             if (arg instanceof COSFloat
-                    && (((COSFloat) arg).doubleValue() > MAX_POSITIVE_FLOAT || ((COSFloat) arg).doubleValue() < MAX_NEGATIVE_FLOAT))
+                    && (((COSFloat) arg).floatValue() > MAX_POSITIVE_FLOAT
+                            || ((COSFloat) arg).floatValue() < MAX_NEGATIVE_FLOAT))
             {
                 throw createLimitError(ERROR_SYNTAX_NUMERIC_RANGE, "Invalid float range in a Number operand");
             }
