+++ preflight/src/main/java/org/apache/pdfbox/preflight/parser/PreflightParser.java
@@ -681,7 +681,7 @@ public class PreflightParser extends PDFParser
             COSNumber number = (COSNumber) result;
             if (number instanceof COSFloat)
             {
-                Double real = number.doubleValue();
+                Float real = number.floatValue();
                 if (real > MAX_POSITIVE_FLOAT || real < MAX_NEGATIVE_FLOAT)
                 {
                     addValidationError(new ValidationError(ERROR_SYNTAX_NUMERIC_RANGE,
